netmiko-lab/requirements.txt
bcrypt==4.0.1
cffi==1.15.1
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
PyNaCl==1.5.0
PyYAML==6.0
